requirements.txt
langchain
PyPDF2
python-dotenv
streamlit
openai
faiss-cpu
altair
tiktoken
huggingface-hub
InstructorEmbedding
sentence-transformers
langdetect